wedding planner project
Wedding planners are well versed in the details of the wedding industry. They have a keen eye for contracts and working with vendors, taking away the guesswork in decision-making and for advising on what often works and what doesn’t. A wedding designer, often confused with a wedding planner brings an artistic eye to the table: